scripts/requirements.txt
paramiko==2.7.2
sshtunnel==0.1.5
pyyaml
